@@ -17,7 +17,7 @@
         "source": "https://github.com/php-tuf/composer-stager"
     },
     "require": {
-        "php": "^7.3 || ^8.0",
+        "php": "^7.4 || ^8.0",
         "symfony/filesystem": "^4.4 || ^5.2 || ^6.1",
         "symfony/process": "^4.4 || ^5.2 || ^6.1"
     },
