@@ -6,6 +6,7 @@ use PhpTuf\ComposerStager\Domain\Aggregate\PreconditionsTree\CommitterPreconditi
 use PhpTuf\ComposerStager\Domain\Exception\ExceptionInterface;
 use PhpTuf\ComposerStager\Domain\Exception\RuntimeException;
 use PhpTuf\ComposerStager\Domain\Service\FileSyncer\FileSyncerInterface;
+use PhpTuf\ComposerStager\Domain\Service\Precondition\PreconditionInterface;
 use PhpTuf\ComposerStager\Domain\Service\ProcessOutputCallback\ProcessOutputCallbackInterface;
 use PhpTuf\ComposerStager\Domain\Service\ProcessRunner\ProcessRunnerInterface;
 use PhpTuf\ComposerStager\Domain\Value\Path\PathInterface;
@@ -13,11 +14,9 @@ use PhpTuf\ComposerStager\Domain\Value\PathList\PathListInterface;
 
 final class Committer implements CommitterInterface
 {
-    /** @var \PhpTuf\ComposerStager\Domain\Service\FileSyncer\FileSyncerInterface */
-    private $fileSyncer;
+    private FileSyncerInterface $fileSyncer;
 
-    /** @var \PhpTuf\ComposerStager\Domain\Service\Precondition\PreconditionInterface */
-    private $preconditions;
+    private PreconditionInterface $preconditions;
 
     public function __construct(CommitterPreconditionsInterface $preconditions, FileSyncerInterface $fileSyncer)
     {
