@@ -117,9 +117,7 @@ final class PhpFileSyncerUnitTest extends TestCase
     {
         // Expose private method for testing.
         // @see https://ocramius.github.io/blog/accessing-private-php-class-members-without-reflection/
-        $method = static function (PhpFileSyncer $sut, $ancestor, $path): string {
-            return $sut::getRelativePath($ancestor, $path);
-        };
+        $method = static fn (PhpFileSyncer $sut, $ancestor, $path): string => $sut::getRelativePath($ancestor, $path);
         $sut = $this->createSut();
         $method = Closure::bind($method, null, $sut);
 
